Fix stack-chain issue
